\name{add_pseudo_counts}
\alias{add_pseudo_counts}
\title{ Add pseudo counts to amino acid matrix based on defined groups }
\description{
  This function will add pseudo counts to binary amino acid matrix based on the defined groups.
  It is used to minimise the effect of small sample size. The method of Henikoff and Henikoff is used to calculate the pseudocounts
  An alternative method is to simply add 1 to the binary matrix
}
\usage{
 add_pseudo_counts(amino,groups)
}
\arguments{
  \item{amino}{ Matrix representation of alignment generated by convert\_aln\_amino }
  \item{groups}{ Vector or factor that shows the group representation for each sequence in the alignment}
}
\examples{
library(bgafun)
data(LDH.amino.gapless)
data(LDH.groups)
LDH.pseudo=LDH.amino.gapless+1
#or use the function 
LDH.pseudo=add_pseudo_counts(LDH.amino.gapless,LDH.groups)
}

\keyword{ manip }
