\name{amino_counts}
\alias{amino_counts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ calculate count of amino acid types at each position }
\description{
 Internal Function
 Calculate the counts of amino acid types at each position in an alignment from a binary amino acid matrix 
}
\keyword{ manip }% at least one, from doc/KEYWORDS