\name{average_cols_aap}
\alias{average_cols_aap}
\title{ Replaces gaps with the average of the column }
\description{
  This function will deal with gaps in the Amino Acid Property encoding scheme
  It replaces gaps with the average in the column for each group, provided the column is highly occupied for that group. 
  It will only average out over columns that have high percentage of gaps
  It will remove all other columns containing gaps.
}
\usage{
 average_cols_aap(x,y)
}
\arguments{
  \item{x}{ Matrix representation of alignment generated by convert\_aln\_AAP }
  \item{y}{ Vector or factor that shows the group representation for each sequence in the alignment}
}
\examples{
library(bgafun)
data(LDH)
data(LDH.groups)
LDH.aap=convert_aln_AAP(LDH)
LDH.aap.ave=average_cols_aap(LDH.aap,LDH.groups)
dim(LDH.aap.ave)
}

\keyword{ IO }
