\name{Calculate_Row_Weights}
\alias{Calculate_Row_Weights}
\title{ Calculate the sequence weights for all the rows in my amino,using label as the grouping }
\description{
 This will calculate the sequence weights for each group using the Heinkoff and Heinkoff method. 
 Each residue in the sequence is assigned a weight depending on how unique it is in the column.
 The sequence weight is then the sum of these weights, and the total weight is the number of groups
}
\usage{
 Calculate_Row_Weights(my_amino,label)
}
\arguments{
 \item{my_amino}{ Matrix representation of alignment generated by convert\_aln\_amino }
  \item{label}{ Vector or factor that shows the group representation for each sequence in the alignment}
}
\references{
 Henikoff, S. and J. G. Henikoff (1994). "Position-based sequence weights." J Mol Biol 243(4): 574-8.
}
\examples{
library("bgafun")
data(LDH.amino.gapless)
data(LDH.groups)
LDH.weights=Calculate_Row_Weights(LDH.amino.gapless,LDH.groups)
sum(LDH.weights)
}


\keyword{ manip }
