\name{convert_amino-package}
\alias{convert_amino-package}
\docType{package}
\title{
The functions required to convert an alignment into a binary matrix suitable for BGA analysis 
}
\description{
The functions required to convert an alignment into a binary matrix suitable for BGA analysis 
}
\details{
Read in the alignment, then convert into matrix
}
\author{
Iain Wallace
}
\references{
BMC hopefully
}
\keyword{ package }
