\name{convertAAP-package}
\alias{convertAAP-package}
\alias{convertAAP}
\docType{package}
\title{
Converts an alignment into a matrix using the AAP encoding 
}
\description{
Convert an alignment read in by seqinr into a matrix using the AAP encoding. this is suitable for BGA analysis using PCA
}
\details{
\tabular{ll}{
Package: \tab convertAAP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-03-14\cr
License: \tab Artistic License\cr
}
}
\author{
Iain Wallace
}
\references{
BMC hopefully
}
\keyword{ package }

