\name{create_probab}
\alias{create_probab}

\title{ Generates probability matrix for pseudocounts calculation}
\description{
  Internal function.
  Generates an amino acid probability matrix which is based on BLOSUM 62, and is used to calculate how many pseudo counts should be added
}
\keyword{ manip }
