\name{create_profile}
\alias{create_profile}

\title{ Creates a sequence profile for an binary amino acid matrix }
\description{        
 Internal Function
 Returns a profile matrix, which show how many of each type of amino acids are in each position in an alignmnet
 It takes in a binary amino acid matrix
}             
\keyword{ manip }
