\name{Henikoff_weights}
\alias{Henikoff_weights}
\title{ Calculates Henikoff weights for each sequence in a binary amino acid matrix }
\description{
 Internal Function
 Calculates a sequence weight for each sequence in an alignment using the Henikoff method.
}
\references{
 Henikoff, S. and J. G. Henikoff (1994). "Position-based sequence weights." J Mol Biol 243(4): 574-8.
}

\keyword{ manip }
