\name{remove_gaps_groups}
\alias{remove_gaps_groups}
\title{ remove gaps from a binary amino matrix }
\description{
 This function is used to deal with gaps in the binary amino acid encoding. 
 It will remove positions from a binary amino matrix that contain more a certain fraction of gaps for any group in a column, in the alignment
 The gap fraction should be between 0 and 1, and can be changed with the gap\_fraction variable.
}
\usage{
 remove_gaps_groups(x,z,gap_fraction=0.6)
}
\arguments{
  \item{x}{ Matrix representation of alignment generated by convert\_aln\_amino }
  \item{z}{ Vector or factor that shows the group representation for each sequence in the alignment}
  \item{gap_fraction}{Float between 0 and 1 indicating the fraction of gaps in a column before it should be removed}
}
\examples{
library(bgafun)
data(LDH)
data(LDH.groups)
LDH.amino=convert_aln_amino(LDH)
dim(LDH.amino)
LDH.amino.gapless=remove_gaps_groups(LDH.amino,LDH.groups,gap_fraction=0.6)
dim(LDH.amino.gapless)
}
\keyword{ manip }
