\name{run_between_pca}
\alias{run_between_pca}
\title{ run PCA to identify functional positions in an alignment }
\description{
 This is a cover function that runs supervised PCA on a matrix that represents an alignment.
 The matrix can either be a binary matrix (with or without pseudocounts) or one that represents the properties at each position of the alignment
}
\usage{
 run_between_pca(x,z,y)
}
\arguments{
  \item{x}{ Matrix representation of alignment generated by convert\_aln\_amino }
  \item{z}{ Matrix representation of alignment generated by convert\_aln\_amino or convert\_aln\_AAP }
  \item{y}{ Vector or factor that shows the group representation for each sequence in the alignment}
}
\examples{
library(bgafun)
data(LDH)
data(LDH.groups)
#Used to calculate the sequence weights
data(LDH.amino.gapless)
data(LDH.aap.ave)
#Run the analysis
LDH.aap.ave.bga=run_between_pca(LDH.amino.gapless,LDH.aap.ave,LDH.groups)
class(LDH.aap.ave.bga)
#to visualise the results
plot(LDH.aap.ave.bga)
}

                                
\keyword{ manip }
