\name{sum_20_aln}
\alias{sum_20_aln}
\title{ Calculates number of amino acids in each group of 20 columns (1 column in an alignment) }
\description{
Internal Function
Calculates number of amino acids in each group of 20 columns  which corresponds to 1 column in an alignment
It takes in an binary amino acid matrix.
}
\keyword{ manip }
