\name{sum_aln}
\alias{sum_aln}
\title{ Calculate number of amino acids in each position in an alignment}
\description{
 Internal Function
 Calculates the total number of amino acids in each position. It is used to identify positions with a high percentage of gaps
 It works on an amino acid matrix
}

\keyword{ manip }