\name{Weight_Amino}
\alias{Weight_Amino}
\title{ Calculates sequence weight for each sequence in an amino acid matrix }
\description{
  Internal Function
  Calculates sequence weight for each sequence, and multiples the matrix by this weight. 
  It returns a weighted amino acid matrix.
}
\keyword{ manip } 
